+++ crates/rust-client/src/account/mod.rs
@@ -35,12 +35,18 @@
 //!
 //! For more details on accounts, refer to the [Account] documentation.
 
-use alloc::vec::Vec;
+use alloc::{string::ToString, vec::Vec};
 
-use miden_objects::{crypto::rand::FeltRng, Word};
+use miden_lib::account::{auth::RpoFalcon512, wallets::BasicWallet};
+use miden_objects::{
+    block::BlockHeader,
+    crypto::{dsa::rpo_falcon512::PublicKey, rand::FeltRng},
+    AccountError, Word,
+};
 
 use super::Client;
 use crate::{
+    rpc::domain::account::AccountDetails,
     store::{AccountRecord, AccountStatus},
     ClientError,
 };
@@ -157,6 +163,27 @@ impl<R: FeltRng> Client<R> {
         }
     }
 
+    /// Imports an account from the network to the client's store. The account needs to be public
+    /// and be tracked by the network, it will be fetched by its ID. If the account was already
+    /// being tracked by the client, it's state will be overwritten.
+    ///
+    /// # Errors
+    /// - If the account is not found on the network.
+    /// - If the account is private.
+    /// - There was an error sending the request to the network.
+    pub async fn import_account_by_id(&mut self, account_id: AccountId) -> Result<(), ClientError> {
+        let account_details = self.rpc_api.get_account_update(account_id).await?;
+
+        let account = match account_details {
+            AccountDetails::Private(..) => {
+                return Err(ClientError::AccountIsPrivate(account_id));
+            },
+            AccountDetails::Public(account, ..) => account,
+        };
+
+        self.add_account(&account, None, true).await
+    }
+
     // ACCOUNT DATA RETRIEVAL
     // --------------------------------------------------------------------------------------------
 
@@ -220,6 +247,57 @@ impl<R: FeltRng> Client<R> {
             .await?
             .ok_or(ClientError::AccountDataNotFound(account_id))
     }
+}
+
+// UTILITY FUNCTIONS
+// ================================================================================================
+
+/// Builds an regular account ID from the provided parameters. The ID may be used along
+/// `Client::import_account_by_id` to import a public account from the network (provided that the
+/// used seed is known).
+///
+/// This function will only work for accounts with the [`BasicWallet`] and [`RpoFalcon512`]
+/// components.
+///
+/// # Arguments
+/// - `init_seed`: Initial seed used to create the account. This is the seed passed to
+///   [`AccountBuilder::new`].
+/// - `public_key`: Public key of the account used in the [`RpoFalcon512`] component.
+/// - `storage_mode`: Storage mode of the account.
+/// - `is_mutable`: Whether the account is mutable or not.
+/// - `anchor_block`: Anchor block of the account.
+///
+/// # Errors
+/// - If the provided block header is not an anchor block.
+/// - If the account cannot be built.
+pub fn build_wallet_id(
+    init_seed: [u8; 32],
+    public_key: PublicKey,
+    storage_mode: AccountStorageMode,
+    is_mutable: bool,
+    anchor_block: BlockHeader,
+) -> Result<AccountId, ClientError> {
+    let account_type = if is_mutable {
+        AccountType::RegularAccountUpdatableCode
+    } else {
+        AccountType::RegularAccountImmutableCode
+    };
+
+    let accound_id_anchor = (&anchor_block).try_into().map_err(|_| {
+        ClientError::AccountError(AccountError::AssumptionViolated(
+            "Provided block header is not an anchor block".to_string(),
+        ))
+    })?;
+
+    let (account, _) = AccountBuilder::new(init_seed)
+        .anchor(accound_id_anchor)
+        .account_type(account_type)
+        .storage_mode(storage_mode)
+        .with_component(RpoFalcon512::new(public_key))
+        .with_component(BasicWallet)
+        .build()?;
+
+    Ok(account.id())
 }
 
 // TESTS