+++ crates/rust-client/src/transaction/mod.rs
@@ -48,7 +48,7 @@
 //!         NoteType::Private,
 //!         client.rng(),
 //!     )?
-//!     .build();
+//!     .build()?;
 //!
 //!     // Execute the transaction. This returns a TransactionResult.
 //!     let tx_result: TransactionResult = client.new_transaction(sender_id, tx_request).await?;
@@ -1094,7 +1094,8 @@ mod test {
             client.rng(),
         )
         .unwrap()
-        .build();
+        .build()
+        .unwrap();
 
         let tx_result = client.new_transaction(account.id(), tx_request).await.unwrap();
         assert!(tx_result