+++ tests/integration/common.rs
@@ -91,7 +91,10 @@ pub fn get_client_config() -> (Endpoint, u64, PathBuf, PathBuf) {
 
     let timeout_ms = rpc_config_toml["timeout"].as_integer().unwrap() as u64;
 
-    (endpoint, timeout_ms, create_test_store_path(), temp_dir())
+    let auth_path = temp_dir().join(format!("keystore-{}", Uuid::new_v4()));
+    std::fs::create_dir_all(&auth_path).unwrap();
+
+    (endpoint, timeout_ms, create_test_store_path(), auth_path)
 }
 
 pub fn create_test_store_path() -> std::path::PathBuf {
@@ -105,13 +108,22 @@ pub async fn insert_new_wallet<R: FeltRng>(
     storage_mode: AccountStorageMode,
     keystore: &FilesystemKeyStore,
 ) -> Result<(Account, Word, SecretKey), ClientError> {
+    let mut init_seed = [0u8; 32];
+    client.rng().fill_bytes(&mut init_seed);
+
+    insert_new_wallet_with_seed(client, storage_mode, keystore, init_seed).await
+}
+
+pub async fn insert_new_wallet_with_seed<R: FeltRng>(
+    client: &mut Client<R>,
+    storage_mode: AccountStorageMode,
+    keystore: &FilesystemKeyStore,
+    init_seed: [u8; 32],
+) -> Result<(Account, Word, SecretKey), ClientError> {
     let key_pair = SecretKey::with_rng(client.rng());
     let pub_key = key_pair.public_key();
 
     keystore.add_key(&AuthSecretKey::RpoFalcon512(key_pair.clone())).unwrap();
-
-    let mut init_seed = [0u8; 32];
-    client.rng().fill_bytes(&mut init_seed);
 
     let anchor_block = client.get_latest_epoch_block().await.unwrap();
 
@@ -320,7 +332,8 @@ pub async fn mint_note(
         client.rng(),
     )
     .unwrap()
-    .build();
+    .build()
+    .unwrap();
     execute_tx_and_sync(client, fungible_asset.faucet_id(), tx_request.clone()).await;
 
     // Check that note is committed and return it
@@ -340,7 +353,8 @@ pub async fn consume_notes(
     println!("Consuming Note...");
     let tx_request =
         TransactionRequestBuilder::consume_notes(input_notes.iter().map(|n| n.id()).collect())
-            .build();
+            .build()
+            .unwrap();
     execute_tx_and_sync(client, account_id, tx_request).await;
 }
 
@@ -372,7 +386,9 @@ pub async fn assert_note_cannot_be_consumed_twice(
     println!("Consuming Note...");
 
     // Double-spend error expected to be received since we are consuming the same note
-    let tx_request = TransactionRequestBuilder::consume_notes(vec![note_to_consume_id]).build();
+    let tx_request = TransactionRequestBuilder::consume_notes(vec![note_to_consume_id])
+        .build()
+        .unwrap();
     match client.new_transaction(consuming_account_id, tx_request).await {
         Err(ClientError::TransactionExecutorError(
             TransactionExecutorError::FetchTransactionInputsFailed(
@@ -407,5 +423,5 @@ pub fn mint_multiple_fungible_asset(
         })
         .collect::<Vec<OutputNote>>();
 
-    TransactionRequestBuilder::new().with_own_output_notes(notes).unwrap().build()
+    TransactionRequestBuilder::new().with_own_output_notes(notes).build().unwrap()
 }