+++ CHANGELOG.md
@@ -2,10 +2,15 @@
 
 ## 0.8.0 (TBD)
 
+### Features
+
+* Added support to import public accounts to `Client` (#733).
+
 ### Changes
 
 * Add check for empty pay to id notes (#714).
 * [BREAKING] Refactored authentication out of the `Client` and added new separate authenticators (#718).
+* Move error handling to the `TransactionRequestBuilder::build()` (#750).
 
 ## 0.7.0 (2025-01-28)